change Track.last_chapter_read to Float
each TrackService can convert it to Int if decimal chapters are not supported


(cherry picked from commit 25ba20cd2a93b66885b7d00306dd331b024ba2b9)

+++ app/src/main/java/eu/kanade/tachiyomi/data/backup/full/models/BackupTracking.kt
@@ -33,7 +33,7 @@ data class BackupTracking(
             library_id = this@BackupTracking.libraryId
             title = this@BackupTracking.title
             // convert from float to int because of 1.x types
-            last_chapter_read = this@BackupTracking.lastChapterRead.toInt()
+            last_chapter_read = this@BackupTracking.lastChapterRead
             total_chapters = this@BackupTracking.totalChapters
             score = this@BackupTracking.score
             status = this@BackupTracking.status
@@ -52,7 +52,7 @@ data class BackupTracking(
                 libraryId = track.library_id!!,
                 title = track.title,
                 // convert to float for 1.x
-                lastChapterRead = track.last_chapter_read.toFloat(),
+                lastChapterRead = track.last_chapter_read,
                 totalChapters = track.total_chapters,
                 score = track.score,
                 status = track.status,
